# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API Documentation",
        "description": "News API Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get list of news",
                "description": "Returns list of news",
                "operationId": "getNewsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "News Post"
                ],
                "summary": "Store new news",
                "description": "Returns news data",
                "operationId": "storeNews",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "news_category_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "short_description": {
                                        "type": "string"
                                    },
                                    "full_content": {
                                        "type": "string"
                                    },
                                    "thumbnail": {
                                        "type": "string"
                                    },
                                    "thumbnail_caption": {
                                        "type": "string"
                                    },
                                    "is_publish": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news by id",
                "description": "Returns news by id",
                "operationId": "getNews",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "news id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "News Update"
                ],
                "summary": "Update news",
                "description": "Returns news data",
                "operationId": "updateNews",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "news id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "news_category_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "short_description": {
                                        "type": "string"
                                    },
                                    "full_content": {
                                        "type": "string"
                                    },
                                    "thumbnail": {
                                        "type": "string"
                                    },
                                    "thumbnail_caption": {
                                        "type": "string"
                                    },
                                    "is_publish": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "News Delete"
                ],
                "summary": "Delete existing news",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteNews",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "news id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "News",
            "description": "API Endpoints of News"
        }
    ],
    "security": [
        []
    ]
}